# cmd/settlr/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Settlr API",
    "version": "1.0.0",
    "description": "Free/open-source Splitwise alternative. Integer cents ledger, Argon2id, JWT refresh rotation, cursor pagination."
  },
  "servers": [
    {
      "url": "https://settlrapi.theswissknife.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/groups/{id}/settlements": {
      "get": {
        "summary": "List settlements",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create settlement",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/settlements/{id}": {
      "patch": {
        "summary": "Update settlement",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete settlement",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/groups/{id}/stats": {
      "get": {
        "summary": "Stats",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}